adapter utils
