+++ src/component/utils/index.ts
@@ -1,4 +1,8 @@
 import { checkDatasource } from './checkDatasource';
 import { assignSettings, assignDevSettings } from './assignSettings';
+import { itemAdapterEmpty, generateMockAdapter, protectAdapterPublicMethod } from './adapter';
 
-export { checkDatasource, assignSettings, assignDevSettings };
+export {
+  checkDatasource, assignSettings, assignDevSettings,
+  itemAdapterEmpty, generateMockAdapter, protectAdapterPublicMethod
+};
